Changed doxygen to generate a toggle for generating docs.
Previously defaulted to AUTO_LIGHT, which would default to light unless
there was a user preference, which my dev environment had as dark.
Now GitHub Actions has no preference so it defaulted to light.

Toggle is a better solution however, as it allows for user choice.

This also tests if it'll upload properly to the server to refresh the
site on new pushes.